Updated Transformer model to latest Swift For TensorFlow build (April 16, 2019).

+++ Transformer/PythonCheckpointReader.swift
@@ -83,7 +83,7 @@ extension MultiHeadAttention: InitializableFromPythonCheckpoint {
         wo = TimeDistributed(Dense<Float>(
             contentsOfPythonCheckpointFile: path,
             scope: scope + "/c_proj"))
-        headCount = Int32(12)
+        headCount = 12
     }
 }
